Extract ParallelFragmentCoordinator into its own module
Move ParallelFragmentCoordinator and _scoped_ctx_attach from fragment.py
into parallel_coordinator.py to break the import cycle between fragment.py
and script_run_context.py. The deferred import in reset() remains since
the two smaller modules still have a mutual dependency, but the cycle no
longer routes through the large fragment module.
